+++ AppCenter/AppCenter/AppCenter.h
@@ -3,11 +3,31 @@
 
 #import <Foundation/Foundation.h>
 
+#if __has_include(<AppCenter/MSACAppCenter.h>)
+#import <AppCenter/MSACAbstractLog.h>
+#import <AppCenter/MSACAppCenter.h>
+#import <AppCenter/MSACAppCenterErrors.h>
+#import <AppCenter/MSACChannelGroupProtocol.h>
+#import <AppCenter/MSACChannelProtocol.h>
+#import <AppCenter/MSACConstants+Flags.h>
+#import <AppCenter/MSACConstants.h>
+#import <AppCenter/MSACCustomProperties.h>
+#import <AppCenter/MSACDevice.h>
+#import <AppCenter/MSACEnable.h>
+#import <AppCenter/MSACLog.h>
+#import <AppCenter/MSACLogWithProperties.h>
+#import <AppCenter/MSACLogger.h>
+#import <AppCenter/MSACService.h>
+#import <AppCenter/MSACServiceAbstract.h>
+#import <AppCenter/MSACWrapperLogger.h>
+#import <AppCenter/MSACWrapperSdk.h>
+#else
 #import "MSACAbstractLog.h"
 #import "MSACAppCenter.h"
 #import "MSACAppCenterErrors.h"
 #import "MSACChannelGroupProtocol.h"
 #import "MSACChannelProtocol.h"
+#import "MSACConstants+Flags.h"
 #import "MSACConstants.h"
 #import "MSACCustomProperties.h"
 #import "MSACDevice.h"
@@ -19,3 +39,4 @@
 #import "MSACServiceAbstract.h"
 #import "MSACWrapperLogger.h"
 #import "MSACWrapperSdk.h"
+#endif