+++ AppCenter/AppCenter/Internals/MSACServiceCommon.h
@@ -3,7 +3,11 @@
 
 #import <Foundation/Foundation.h>
 
+#if __has_include(<AppCenter/MSACConstants.h>)
+#import <AppCenter/MSACConstants.h>
+#else
 #import "MSACConstants.h"
+#endif
 
 NS_ASSUME_NONNULL_BEGIN