+++ AppCenterAnalytics/AppCenterAnalytics/AppCenterAnalytics.h
@@ -3,11 +3,18 @@
 
 #import <Foundation/Foundation.h>
 
-#import "MSACAbstractLog.h"
+#if __has_include(<AppCenterAnalytics/MSACAnalytics.h>)
+#import <AppCenterAnalytics/MSACAnalytics.h>
+#import <AppCenterAnalytics/MSACAnalyticsAuthenticationProvider.h>
+#import <AppCenterAnalytics/MSACAnalyticsAuthenticationProviderDelegate.h>
+#import <AppCenterAnalytics/MSACAnalyticsTransmissionTarget.h>
+#import <AppCenterAnalytics/MSACEventLog.h>
+#import <AppCenterAnalytics/MSACEventProperties.h>
+#else
 #import "MSACAnalytics.h"
 #import "MSACAnalyticsAuthenticationProvider.h"
 #import "MSACAnalyticsAuthenticationProviderDelegate.h"
 #import "MSACAnalyticsTransmissionTarget.h"
-#import "MSACConstants+Flags.h"
 #import "MSACEventLog.h"
 #import "MSACEventProperties.h"
+#endif