+++ AppCenterAnalytics/AppCenterAnalytics/Model/MSACEventLog.h
@@ -1,7 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+#ifndef MSAC_EVENT_LOG_H
+#define MSAC_EVENT_LOG_H
+
+#if __has_include(<AppCenterAnalytics/MSACLogWithNameAndProperties.h>)
+#import <AppCenterAnalytics/MSACLogWithNameAndProperties.h>
+#else
 #import "MSACLogWithNameAndProperties.h"
+#endif
 
 @class MSACEventProperties;
 @class MSACMetadataExtension;
@@ -20,3 +27,5 @@ NS_SWIFT_NAME(EventLog)
 @property(nonatomic, strong) MSACEventProperties *typedProperties;
 
 @end
+
+#endif