+++ AppCenterDistribute/AppCenterDistribute/MSACDistributeUtil.m
@@ -13,16 +13,23 @@ NSBundle *MSACDistributeBundle(void) {
   static NSBundle *bundle = nil;
   static dispatch_once_t predicate;
   dispatch_once(&predicate, ^{
-    // The resource bundle is part of the main app bundle, e.g. .../Puppet.app/AppCenterDistribute.bundle
-    NSString *mainBundlePath = [[NSBundle bundleForClass:[MSACDistribute class]] resourcePath];
-    NSString *frameworkBundlePath = [mainBundlePath stringByAppendingPathComponent:APP_CENTER_DISTRIBUTE_BUNDLE];
-    bundle = [NSBundle bundleWithPath:frameworkBundlePath];
+
+  // The resource bundle is part of the main app bundle, e.g. .../Puppet.app/AppCenterDistribute.bundle
+#ifdef SWIFTPM_MODULE_BUNDLE
+    bundle = SWIFTPM_MODULE_BUNDLE;
+#else
+    NSBundle *mainBundle = [NSBundle bundleForClass:[MSACDistribute class]];
+    NSURL *url = [mainBundle URLForResource:APP_CENTER_DISTRIBUTE_BUNDLE_NAME withExtension:@"bundle"];
+    if (url) {
+      bundle = [NSBundle bundleWithURL:url];
+    }
 
     // Log to console in case the bundle is nil.
     if (!bundle) {
       MSACLogError([MSACDistribute logTag], @"The AppCenterDistributeResources.bundle file could not be found in your app. "
                                             @"Please add it to your project as described in our readme.");
     }
+#endif
   });
   return bundle;
 }