+++ Package@swift-5.3.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.0
+// swift-tools-version:5.3
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
@@ -7,6 +7,7 @@ import PackageDescription
 
 let package = Package(
     name: "AppCenter",
+    defaultLocalization: "en",
     platforms: [
         .iOS(.v9),
         .macOS(.v10_10),
@@ -20,10 +21,14 @@ let package = Package(
         .library(
             name: "AppCenterCrashes",
             type: .static,
-            targets: ["AppCenterCrashes"])
+            targets: ["AppCenterCrashes"]),
+        .library(
+            name: "AppCenterDistribute",
+            type: .static,
+            targets: ["AppCenterDistribute"]),
     ],
     dependencies: [
-        .package(url: "https://github.com/microsoft/plcrashreporter.git", .upToNextMinor(from: "1.8.0")),
+        .package(url: "https://github.com/microsoft/PLCrashReporter.git", .upToNextMinor(from: "1.8.0")),
     ],
     targets: [
         .target(
@@ -31,7 +36,7 @@ let package = Package(
             path: "AppCenter/AppCenter",
             exclude: ["Support"],
             cSettings: [
-                .define("APP_CENTER_C_VERSION", to:"\"4.1.0\""),
+                .define("APP_CENTER_C_VERSION", to:"\"4.0.1\""),
                 .define("APP_CENTER_C_BUILD", to:"\"1\""),
                 .headerSearchPath("**"),
             ],
@@ -62,9 +67,12 @@ let package = Package(
         ),
         .target(
             name: "AppCenterCrashes",
-            dependencies: ["AppCenter", "CrashReporter"],
+            dependencies: [
+                "AppCenter",
+                .product(name: "CrashReporter", package: "PLCrashReporter"),
+            ],
             path: "AppCenterCrashes/AppCenterCrashes",
-            exclude: ["Support"],
+            exclude: ["Support", "Internals/MSACCrashesBufferedLog.hpp"],
             cSettings: [
                 .headerSearchPath("**"),
                 .headerSearchPath("../../AppCenter/AppCenter/**"),
@@ -73,6 +81,24 @@ let package = Package(
                 .linkedFramework("Foundation"),
                 .linkedFramework("UIKit", .when(platforms: [.iOS, .tvOS])),
                 .linkedFramework("AppKit", .when(platforms: [.macOS])),
+            ]
+        ),
+        .target(
+            name: "AppCenterDistribute",
+            dependencies: ["AppCenter"],
+            path: "AppCenterDistribute/AppCenterDistribute",
+            exclude: ["Support"],
+            resources: [
+                .process("Resources/AppCenterDistribute.strings"),
+            ],
+            cSettings: [
+                .headerSearchPath("**"),
+                .headerSearchPath("../../AppCenter/AppCenter/**"),
+            ],
+            linkerSettings: [
+                .linkedFramework("Foundation"),
+                .linkedFramework("SafariServices", .when(platforms: [.iOS])),
+                .linkedFramework("UIKit", .when(platforms: [.iOS])),
             ]
         )
     ]