+++ Sasquatch/SasquatchSwift/AppDelegate.swift
@@ -315,6 +315,27 @@ extension AppDelegate: DistributeDelegate {
     }
     return false
   }
+  
+  func distributeNoReleaseAvailable(_ distribute: Distribute) {
+    NSLog("distributeNoReleaseAvailable invoked");
+    let alert = UIAlertController(title: nil, message: "No updates available", preferredStyle: .alert)
+    alert.addAction(UIAlertAction(title: "OK", style: .default, handler: nil))
+    self.window?.rootViewController?.present(alert, animated: true)
+  }
+
+  func distributeWillExitApp(_ distribute: Distribute) {
+    print("distributeWillExitApp callback invoked");
+    DispatchQueue.main.async {
+      let alertController = UIAlertController(title: NSLocalizedString("distribute_alert_willExit_title", tableName: "Sasquatch", comment: ""),
+              message: NSLocalizedString("distribute_alert_willExit_message", tableName: "Sasquatch", comment: ""),
+              preferredStyle: .alert)
+
+      // Show alert saying that the app is closing
+      self.window?.rootViewController?.present(alertController, animated: true)
+    }
+
+    sleep(7)
+  }
 }
 
 #endif